Update picksobjectcodeVer2.py
this is the code that worked with scott but not with mine. the version of the code that i'm working with deviates on line 24. that is the path to the pick file on scott's computer. my code has that replaced with the path to the pick file on m machine.